Check length when comparing in refreshOptions
If we don't check length in compareObjects, we won't notice when the new
options object has had a new property added. (Because compareObjects
otherways only checks that everything present in the old options exist
in the new, not the other way round.) This might be especially relevant
when you want to activate an extension like reorderableRows.