add iconSizeClass
IconSize is not changing according to tableGrid IconSize option. So, I added an extra option "exportIconSize" that allow to define an icon size.
It would be much better, if we managed to inherit the main table iconSize option and have the possibility to override it in the plugin options. Instead of defining an extra option.