fix #2014 - ints break toggleColumn
fix #2014 - toggleColumn (due to getFieldIndex) fails unless column.field is string, and likely any other getFieldIndex calls too.

Force it to be stored as string around line 618 (`initTable`) 

OR fix when this is compared, which appears to only be line 49 (`getFieldIndex`). 

But since a wider search for ".field" showed the second approach might not cover all 'other' quirks if js doesnt natively store this as string, so went with forcing it in `initTable`. I went with `new String(oldval).valueOf()` over something like `"" + oldval` to help answer outside use cases as `new String()` (i assume) is more robust than forced concat.

Changes can be proofed using https://jsfiddle.net/dabros/johm6kd3/2/

---

As a side note, `var column` is misleading as it is actually the array of all columns in the current `<tr>`. Maybe a rename to `rowcols` would make this more obvious.