Add rememberOrder option to table what allows to sort a table with or without remembering the order.
This provides more natural and intuitive behaviour. When the option is set to false,
on a first click on a sort arrow for column that is not currently sorted,
table always gets sorted in the same order, thus only when followed by a second click
it gets sorted in the opposite order.