fixed racecondition
When filling in table-search-input-fields, i allways get missing letters while typing. Thats because of the line edited in this change.
You dont' wanna "resett" the field value. in the case that it is a select field, you wanna set the value, thats okay. But not on an input.
setting the value triggers change, so we'll trigger this manually.