Make pageList "All" option locale independant
With current implementation, having the "All" option in the pageList requires declaring it in the table options with the same locale as the one used at runtime. This might be an issue when the local is chosen dynamically depending on client configuration.

For instance, if locale is fr-FR, pageList = [ 10, 25, all ] will show [10, 25, NaN] in the page list, and pageList = [ 10, 25, tous ] will show a proper list. But pageList = [ 10, 25, tous ] will display [10, 25, NaN] if locale is en-US.

This change allows to have the option in every locale when it is declared in english. pageList = [ 10, 25, all ] will show [10,25,Tous] with fr-FR, [10,25,All] with en-US, and so on.