Fixed cookie in combination with filter-control + strict search
When using cookie in combination with filter-control and strict search enabled on a column, selecting the empty value would store it as an empty value in cookie. When returning the page and thus getting the cookie, it would result in no results in bootstrap-table because it would literally search for ''.