Add: Allow for custom buttons to clear one and/or clear all filters
Provide the ability to set table properties defining custom 'clear all'
and/or 'clear one' jQuery selectors for
elements that will
clear filters on click.  See the updated documentation for usage.

Also fixed some tab levels and prefixed jQuery variables with $
where missing in the functions involved in this commit.