Checked property ignored
The checked property is ignored if the field has a value.

For example:
I am setting the checkbox column to the Id field so there is always a value. If I set the formatter to return { checked: false, disabled: true } the checkbox is still checked because the field has a value.