Indicate URL parameters for server side pagination (#6429)
* Indicate URL parameters for server side pagination

I'd tried to use bootstrap table's server side pagination over a year ago and never figured out how to get it to work.  I was working in Django and tried to implement the view that the data-url was calling, but I was missing this key piece of information.  I happened upon the URL parameters in the last answer of a stack post and it was like an epiphany.  I went back to the documentation to see how I'd possibly missed the URL parameters that the view needed to be able to obtain, and was shocked to see that it wasn't a part of the documentation.  Hopefully, this addition to the documentation will make the server-side pagination attempts of others easier to implement...

And if there are other parameters that could be passed, such as for `serverSort`, it would be good to add those as well.

* Update table-options.md

* Update site/docs/api/table-options.md

* Update site/docs/api/table-options.md

* Update table-options.md

* Update table-options.md

* Update table-options.md

Co-authored-by: 文翼 <wenzhixin2010@gmail.com>