Set `aria-sort` attribute on sortable headers
This will improve accessibility of sortable tables in that screen
readers will be able to dictate the current sort order of columns as the
user navigates through them.  Tested with JAWS.

The attribute should be set to "ascending" or "descending" only on the
column that is currently sorted.

See https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Reference/Attributes/aria-sort
for an explanation of this attribute.